+++ Assets/NowUI/Runtime/NowUIInput.cs
@@ -774,6 +774,7 @@ namespace NowUI
         }
     }
 
+    [NowScope]
     public struct NowUIInputScope : IDisposable
     {
         readonly NowUIInputSurface _previousSurface;